transaction/requirements.txt
tushare>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
python-dotenv>=0.19.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
ta-lib>=0.4.24
akshare>=1.16.60